Modify SimpleCatalogManager for "local" downloads
Allow creating a SimpleCatalogManager with a NULL DownloadManager. In
this case, the manager will perform local file copies instead of CURL downloads.

+++ cvmfs/catalog_mgr_ro.cc
@@ -2,8 +2,8 @@
  * This file is part of the CernVM file system.
  */
 
-#include "cvmfs_config.h"
 #include "catalog_mgr_ro.h"
+#include "cvmfs_config.h"
 
 #include "compression.h"
 #include "download.h"
@@ -21,44 +21,66 @@ namespace catalog {
  *                     name of the catalog afterwards
  * @return 0 on success, different otherwise
  */
-LoadError SimpleCatalogManager::LoadCatalog(const PathString  &mountpoint,
-                                            const shash::Any  &hash,
-                                            std::string       *catalog_path,
-                                            shash::Any        *catalog_hash)
-{
+LoadError SimpleCatalogManager::LoadCatalog(const PathString &mountpoint,
+                                            const shash::Any &hash,
+                                            std::string *catalog_path,
+                                            shash::Any *catalog_hash) {
   shash::Any effective_hash = hash.IsNull() ? base_hash_ : hash;
   assert(shash::kSuffixCatalog == effective_hash.suffix);
   const string url = stratum0_ + "/data/" + effective_hash.MakePath();
-  FILE *fcatalog = CreateTempFile(dir_temp_ + "/catalog", 0666, "w",
-                                  catalog_path);
-  if (!fcatalog) {
+  if (local_only_) {
+    *catalog_path = CreateTempPath(dir_temp_ + "/catalog", 0666);
+    if (zlib::DecompressPath2Path(url, *catalog_path)) {
+      LogCvmfs(kLogCatalog, kLogStderr,
+               "SimpleCatalogManager:: DecompressPath2Path: %s -> %s",
+               url.c_str(), catalog_path->c_str());
+
+      shash::Any new_hash(hash.algorithm);
+      if (!shash::HashFile(*catalog_path, &new_hash)) {
+        LogCvmfs(kLogCatalog, kLogStderr,
+                 "Could not hash decompressed catalog");
+        assert(false);
+      }
+
+      LogCvmfs(kLogCatalog, kLogStderr, "SimpleCatalogManager:: HashFile: %s",
+               catalog_hash->ToString(true).c_str());
+
+      *catalog_hash = new_hash;
+      return kLoadNew;
+    }
     LogCvmfs(kLogCatalog, kLogStderr,
-             "failed to create temp file when loading %s", url.c_str());
+             "SimpleCatalogManager:: Could not decompress catalog");
     assert(false);
+  } else {
+    FILE *fcatalog =
+        CreateTempFile(dir_temp_ + "/catalog", 0666, "w", catalog_path);
+    if (!fcatalog) {
+      LogCvmfs(kLogCatalog, kLogStderr,
+               "failed to create temp file when loading %s", url.c_str());
+      assert(false);
+    }
+
+    download::JobInfo download_catalog(&url, true, false, fcatalog,
+                                       &effective_hash);
+    download::Failures retval = download_manager_->Fetch(&download_catalog);
+    fclose(fcatalog);
+
+    if (retval != download::kFailOk) {
+      LogCvmfs(kLogCatalog, kLogStderr,
+               "failed to load %s from Stratum 0 (%d - %s)", url.c_str(),
+               retval, download::Code2Ascii(retval));
+      unlink(catalog_path->c_str());
+      assert(false);
+    }
+
+    *catalog_hash = effective_hash;
+    return kLoadNew;
   }
-
-  download::JobInfo download_catalog(&url, true, false, fcatalog,
-                                     &effective_hash);
-  download::Failures retval = download_manager_->Fetch(&download_catalog);
-  fclose(fcatalog);
-
-  if (retval != download::kFailOk) {
-    LogCvmfs(kLogCatalog, kLogStderr,
-             "failed to load %s from Stratum 0 (%d - %s)", url.c_str(),
-             retval, download::Code2Ascii(retval));
-    unlink(catalog_path->c_str());
-    assert(false);
-  }
-
-  *catalog_hash = effective_hash;
-  return kLoadNew;
 }
 
-
-Catalog* SimpleCatalogManager::CreateCatalog(const PathString  &mountpoint,
-                                             const shash::Any  &catalog_hash,
-                                             Catalog           *parent_catalog)
-{
+Catalog *SimpleCatalogManager::CreateCatalog(const PathString &mountpoint,
+                                             const shash::Any &catalog_hash,
+                                             Catalog *parent_catalog) {
   Catalog *new_catalog = new Catalog(mountpoint, catalog_hash, parent_catalog);
   if (manage_catalog_files_) {
     new_catalog->TakeDatabaseFileOwnership();
